Controlled all the "er" annotations; they're finished now. Some of them where double, they were already part of an other .eaf file. I did not check those. It was the case for: Batch_0010/201902010_02_4 which was completly in Batch_0007/20190101_02. I skipped that one entirely (from minute 6, when I realized it)
Batch_011/20190402_1 was part of Batch_0008/20190402. But the end of
Batch_0008/20190402 was in Batch_011/20190402_1 so I picked up checking
the annotions from that time on, which was at 01:26.

All batches of batch_0012 where completly in Batch_0009/20190404_02, so
I skipped batch_0012 entirely (from 01:00:37 of batch_0012/20190404_02_1
when I realized it).

Added my conventions